[SHDOCVW][SDK] Initial support of Favorites bar
